requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.42.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
datasets==2.19.0
dill==0.3.8
docker-pycreds==0.4.0
evaluate==0.4.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
peft==0.10.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
responses==0.18.0
safetensors==0.4.3
scipy==1.13.0
sentry-sdk==2.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.6
xxhash==3.4.1
yarl==1.9.4
